app/requirements.txt
Flask == 2.1.2
prometheus_client == 0.14.1
google-cloud-secret-manager